hubble/metrics: Support new metrics context: "labelsContext"
The new "labelsContext" metrics context differs from the existing
contexts in that instead of putting different values into the
source/destination labels, it accepts a list of predefined labels that
should be exposed, and maps the flow data into these different labels.

This makes the hubble metrics much easier to use as it now separates out
the different values into separate labels.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>